MySQL prepared batch support (#335)
* migrate pg preparedBatch test to TCK


* add support for MySQL preparedBatch


* minor
